test: proof working
