+++ tests/scrvusd/oracle/test_set_max_acceleration.py
