+++ tests/scrvusd/oracle/test_set_verifier.py
